+++ changelog.md
@@ -7,7 +7,16 @@
 - updated readme
 - reworked the whole file to use more consistant x and y naming
 
-## Patch 0.0.9
+## Patch 0.0.9 // 8/01/2025 
 
 - added exponential force curves
+
+## Patch 0.0.10 //  8/15/2025
+
+- fixed mini map direction indicator
+
+## Patch 0.0.11 // 8/23/2025
+
+- added bullets (replacing wall remover gun)
+- optional gun overlay